move file handling to util
